Move static files into module
