+++ src-hono/log.ts
@@ -1,0 +1,3 @@
+export default function log(message: string, ...rest: string[]) {
+	console.log(`[${new Date().toISOString()}]`, message, ...rest);
+}
